pyproject.toml
[project]
name = "dj3k"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "django==5.2.7",
    "django-prometheus>=2.4.1",
    "psycopg2-binary==2.9.11",
    "pyrsistent==0.20.0",
    "python-dotenv==1.2.1",
]

[dependency-groups]
dev = [
    "hypothesis[cli,django]~=6.142.4",
    "pre-commit~=4.3.0",
    "pytest~=8.4.2",
    "pytest-django~=4.11.1",
    "ruff~=0.14.2",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.settings"
testpaths = ["tests"]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
preview = true
select = [
    "YTT",
    "ANN",
    "ASYNC",
    "S",
    "BLE",
    "FBT",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "DJ",
    "EM",
    "EXE",
    "FIX001",
    "INT",
    "ISC",
    "ICN",
    "LOG",
    "G",
    "INP",
    "PIE",
    "T20",
    "PYI",
    "PT",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SIM",
    "SLOT",
    "TID",
    "TD",
    "PTH",
    "FLY",
    "I",
    "C90",
    "NPY",
    "PD",
    "N",
    "PERF",
    "E",
    "W",
    "DOC202",
    "DOC403",
    "DOC502",
    "D200",
    "D206",
    "D208",
    "D209",
    "D210",
    "D300",
    "D301",
    "D402",
    "D404",
    "D419",
    "F",
    "PL",
    "UP",
    "FURB",
    "RUF",
    "TRY"
]

exclude = [
    "TD002",
    "TD003",
]

unfixable = ["F401"]

[tool.ruff.format]
docstring-code-format = true
